Moved header files to include/ & Reorganized CMakeLists structure
