+++ src/logger.cpp
@@ -1,4 +1,4 @@
-#include "logger.hpp"
+#include "../include/logger.hpp"
 
 // Private----------------------------------------------------------------------
 std::string Logger::getDate() {
