+++ src/motor.cpp
@@ -1,4 +1,4 @@
-#include "motor.hpp"
+#include "../include/motor.hpp"
