+++ test/actuationPlanTest.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 #include <gtest/gtest.h>
-#include "../src/surfaceFitModel.hpp"
-#include "../src/actuationPlan.hpp"
-#include "../src/rocketUtils.hpp"
+#include "../include/surfaceFitModel.hpp"
+#include "../include/actuationPlan.hpp"
+#include "../include/rocketUtils.hpp"
 
 
 class ActuationPlanTest : public ::testing::Test {
@@ -34,14 +34,13 @@ TEST_F(ActuationPlanTest, runPlan) {
     rocket.altiInitFail = false;
     rocket.altiReadFail = false;
 
-
     // Test when Vehicle Status: Glide
     rocket.fail_time = (time_t)(-1);
     rocket.deploy_time = time(nullptr);
     rocket.status = GLIDE;
     rocket.filtered_altitude = 0;
     rocket.filtered_velocity = 0;
-	plan->runPlan(&rocket);
+	plan->runPlan(rocket);
     EXPECT_NEAR(rocket.deployment_angle, 1, 0.01); // Add true values
     EXPECT_NE(rocket.fail_time, (time_t)(-1));
 
@@ -50,7 +49,7 @@ TEST_F(ActuationPlanTest, runPlan) {
     rocket.status = APOGEE;
     rocket.filtered_altitude = 0;
     rocket.filtered_velocity = 0;
-	plan->runPlan(&rocket);
+	plan->runPlan(rocket);
     EXPECT_NEAR(rocket.deployment_angle, 1, 0.01); // Add true values
 
     // Test when IMU Read fails
