+++ test/loggerTest.cpp
@@ -1,4 +1,4 @@
-#include "../src/logger.hpp"
+#include "../include/logger.hpp"
 
 class Two {
