+++ test/output.txt
@@ -1,0 +1,110 @@
+02:33:06-> [INFO]: Log Start---- (Sat Nov 4 2023)
+02:33:06-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+02:33:06-> [ERROR]: Testing Error lksdj fa 3459374
+02:33:06-> [INFO]: Log End
+
+02:34:05-> [INFO]: Log Start---- (Sat Nov 4 2023)
+02:34:05-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+02:34:05-> [ERROR]: Testing Error lksdj fa 3459374
+02:34:05-> [INFO]: Log End
+
+02:44:43-> [INFO]: Log Start---- (Sat Nov 4 2023)
+02:44:43-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+02:44:43-> [ERROR]: Testing Error lksdj fa 3459374
+02:44:43-> [INFO]: Log End
+
+02:44:54-> [INFO]: Log Start---- (Sat Nov 4 2023)
+02:44:54-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+02:44:54-> [ERROR]: Testing Error lksdj fa 3459374
+02:44:54-> [INFO]: Log End
+
+02:45:05-> [INFO]: Log Start---- (Sat Nov 4 2023)
+02:45:05-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+02:45:05-> [ERROR]: Testing Error lksdj fa 3459374
+02:45:05-> [INFO]: Log End
+
+02:45:55-> [INFO]: Log Start---- (Sat Nov 4 2023)
+02:45:55-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+02:45:55-> [ERROR]: Testing Error lksdj fa 3459374
+02:45:55-> [INFO]: Log End
+
+02:47:51-> [INFO]: Log Start---- (Sat Nov 4 2023)
+02:47:51-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+02:47:51-> [ERROR]: Testing Error lksdj fa 3459374
+02:48:39-> [INFO]: Log Start---- (Sat Nov 4 2023)
+02:48:39-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+02:48:39-> [ERROR]: Testing Error lksdj fa 3459374
+11:12:33-> [INFO]: Log Start---- (Sat Nov 4 2023)
+11:12:33-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+11:12:33-> [ERROR]: Testing Error lksdj fa 3459374
+11:16:49-> [INFO]: Log Start---- (Sat Nov 4 2023)
+11:16:49-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+11:16:49-> [ERROR]: Testing Error lksdj fa 3459374
+11:16:57-> [INFO]: Log Start---- (Sat Nov 4 2023)
+11:16:57-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+11:16:57-> [ERROR]: Testing Error lksdj fa 3459374
+11:20:16-> [INFO]: Log Start---- (Sat Nov 4 2023)
+11:20:16-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+11:20:16-> [ERROR]: Testing Error lksdj fa 3459374
+11:20:16-> [INFO]: Log End
+
+11:23:36-> [INFO]: Log Start---- (Sat Nov 4 2023)
+11:23:36-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+11:23:36-> [ERROR]: Testing Error lksdj fa 3459374
+11:23:36-> [INFO]: Log End
+
+11:27:30-> [INFO]: Log Start---- (Sat Nov 4 2023)
+11:27:30-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+11:27:30-> [ERROR]: Testing Error lksdj fa 3459374
+11:27:30-> [INFO]: Log End
+
+11:28:33-> [INFO]: Log Start---- (Sat Nov 4 2023)
+11:28:33-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+11:28:33-> [ERROR]: Testing Error lksdj fa 3459374
+11:28:33-> [INFO]: Log End----
+
+11:39:43-> [INFO]: Log Start---- (Sat Nov 4 2023)
+11:39:43-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+11:39:43-> [INFO]: Testing Info Just some stuff
+11:39:43-> [INFO]: Testing something words words words
+11:39:43-> [ERROR]: Testing Error lksdj fa 3459374
+11:39:43-> [INFO]: Testing blah blah blah blah blah lbal lab lbas labal abala
+11:39:43-> [INFO]: Testing Info aksdjflkas l11111111111111111111111111111111111111111485 345
+11:39:43-> [ERROR]: Testing Error Some other type of error
+11:39:43-> [ERROR]: Testing Error Another one
+11:39:43-> [INFO]: Log End----
+
+13:33:22-> [INFO]: Log Start---- (Tues Nov 7 2023)
+13:33:22-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+13:33:22-> [INFO]: Testing Info Just some stuff
+13:33:22-> [INFO]: Testing something words words words
+13:33:22-> [ERROR]: Testing Error lksdj fa 3459374
+13:33:22-> [INFO]: Testing blah blah blah blah blah lbal lab lbas labal abala
+13:33:22-> [INFO]: Testing Info aksdjflkas l11111111111111111111111111111111111111111485 345
+13:33:22-> [ERROR]: Testing Error Some other type of error
+13:33:22-> [ERROR]: Testing Error Another one
+13:33:22-> [INFO]: Log End----
+
+13:34:00-> [INFO]: Log Start---- (Tues Nov 7 2023)
+13:34:00-> [INFO]: Testing Info aksdjflkas lksajfdlasjfaowe aslkdjf alskjf asodfj03945430 0349534 5039485 345
+13:34:00-> [INFO]: Testing Info Just some stuff
+13:34:00-> [INFO]: Testing something words words words
+13:34:00-> [ERROR]: Testing Error lksdj fa 3459374
+13:34:00-> [INFO]: Testing blah blah blah blah blah lbal lab lbas labal abala
+13:34:00-> [INFO]: Testing Info aksdjflkas l11111111111111111111111111111111111111111485 345
+13:34:00-> [ERROR]: Testing Error Some other type of error
+13:34:00-> [ERROR]: Testing Error Another one
+13:34:00-> [INFO]: Log End----
+
+13:37:22-> [INFO]: Log Start---- (Tues Nov 7 2023)
+13:37:22-> [INFO]: THIS IS A NEEEEEWWWW TEST!!!!!
+13:39:34-> [INFO]: Log Start---- (Tues Nov 7 2023)
+13:39:34-> [INFO]: THIS IS A NEEEEEWWWW TEST!!!!!
+13:39:34-> [ERROR]: ERRORRRRRRRR
+13:39:34-> [INFO]: Log End----
+
+13:41:04-> [INFO]: Log Start---- (Tues Nov 7 2023)
+13:41:04-> [INFO]: THIS IS A NEEEEEWWWW TEST!!!!!
+13:41:04-> [ERROR]: ERRORRRRRRRR
+13:41:04-> [INFO]: Log End----
+
